backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
firebase-admin>=6.0,<7.0
requests>=2.31,<3.0
